+++ merlin/datasets/ecommerce/aliccp/raw/schema.pbtxt
