+++ merlin/datasets/entertainment/movielens/1m/schema.pbtxt
