+++ merlin/datasets/entertainment/movielens/25m/schema.pbtxt
