--- merlin/models/data/advertising/__init__.py
@@ -1,23 +1,0 @@
-#
-# Copyright (c) 2021, NVIDIA CORPORATION.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-from merlin.models.data.advertising.criteo.dataset import (
-    default_criteo_transform,
-    get_criteo,
-    transform_criteo,
-)
-
-__all__ = ["get_criteo", "transform_criteo", "default_criteo_transform"]
