--- merlin/models/data/ecommerce/__init__.py
@@ -1,24 +1,0 @@
-#
-# Copyright (c) 2021, NVIDIA CORPORATION.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-
-from merlin.models.data.ecommerce.aliccp.dataset import (
-    default_aliccp_transformation,
-    get_aliccp,
-    prepare_alliccp,
-    transform_aliccp,
-)
-
-__all__ = ["prepare_alliccp", "transform_aliccp", "get_aliccp", "default_aliccp_transformation"]
