--- merlin/models/data/entertainment/__init__.py
@@ -1,30 +1,0 @@
-#
-# Copyright (c) 2021, NVIDIA CORPORATION.
-#
-# Licensed under the Apache License, Version 2.0 (the "License");
-# you may not use this file except in compliance with the License.
-# You may obtain a copy of the License at
-#
-#     http://www.apache.org/licenses/LICENSE-2.0
-#
-# Unless required by applicable law or agreed to in writing, software
-# distributed under the License is distributed on an "AS IS" BASIS,
-# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-# See the License for the specific language governing permissions and
-# limitations under the License.
-#
-from merlin.models.data.entertainment.movielens.dataset import (
-    default_ml1m_transformation,
-    default_ml25m_transformation,
-    default_ml100k_transformation,
-    get_movielens,
-    transform_movielens,
-)
-
-__all__ = [
-    "get_movielens",
-    "transform_movielens",
-    "default_ml100k_transformation",
-    "default_ml1m_transformation",
-    "default_ml25m_transformation",
-]
