+++ merlin/models/tf/__init__.py
@@ -27,8 +27,8 @@ from tensorflow.python.keras.metrics import Metric
 from tensorflow.python.keras.optimizer_v2.optimizer_v2 import OptimizerV2
 from tensorflow.python.training.tracking.data_structures import ListWrapper, _DictWrapper
 
-from merlin.models import data
-from merlin.models.data.synthetic import generate_data
+from merlin import datasets
+from merlin.datasets.synthetic import generate_data
 from merlin.models.tf.blocks.core.aggregation import (
     ConcatFeatures,
     ElementwiseSum,
@@ -184,7 +184,7 @@ __all__ = [
     "StochasticSwapNoise",
     "ExpandDims",
     "NoOp",
-    "data",
+    "datasets",
     "generate_data",
     "ItemSampler",
     "EmbeddingWithMetadata",
